added to templating

+++ preprocessing/template_utils.py
@@ -22,18 +22,49 @@ def create_mdps(
     if not os.path.exists(template_path):
         raise FileNotFoundError(f"Template path {template_path} does not exist")
 
-    with open(template_path, "r") as file:
-        content = file.read()
-
-    content = content.replace("{temp}", str(simulation_temp_k))
+    # Construct the output file path
     output_filename = MDP_NAMING_SCHEME.format(
         run_type=mdp_type.value, temp=simulation_temp_k
     )
     output_path = os.path.join(output_dir, output_filename)
+
+    # Check if the file already exists
+    if os.path.exists(output_path):
+        logger.info(f"MDP file already exists at {output_path}, skipping creation.")
+        return output_path
+
+    # Read and modify the template content
+    with open(template_path, "r") as file:
+        content = file.read()
+
+    content = content.replace("{temp}", str(simulation_temp_k))
     os.makedirs(output_dir, exist_ok=True)
 
+    # Write the modified content to the output file
     with open(output_path, "w") as file:
         file.write(content)
 
     logger.info(f"Created MDP file at {output_path}")
+    return output_path
+
+
+def retrieve_mdps(
+    mdp_type: TemplatedMdps,
+    simulation_temp_k: float,
+    searched_dir: Optional[str] = None,
+):
+    if not searched_dir:
+        searched_dir = MDP_DIRS[mdp_type.value]
+
+    output_filename = MDP_NAMING_SCHEME.format(
+        run_type=mdp_type.value, temp=simulation_temp_k
+    )
+    output_path = os.path.join(searched_dir, output_filename)
+
+    if not os.path.exists(output_path):
+        raise FileNotFoundError(
+            f"MDP file {output_path} does not exist, run create_mdps() from template_utils"
+        )
+
+    logger.info(f"Retrieved MDP file from {output_path}")
     return output_path
